[0580] Amended form base
